+++ FM.Application/Command/UserCommand.cs
@@ -23,13 +23,7 @@ namespace FM.Application.Command
             await _unitOfWork.BeginTransactionAsync();
             try
             {
-                var user = new User
-                {
-                    Id = Guid.NewGuid().ToString(), // Set the Id to a new GUID
-                    DisplayName = command.DisplayName,
-                    SpotifyUserId = command.SpotifyUserId,
-                    UserRoleId = command.UserRoleId
-                };
+                var user = new User(Guid.NewGuid().ToString(), command.DisplayName, command.SpotifyUserId, command.UserRoleId);
 
                 await _userRepository.AddUserAsync(user);
                 await _unitOfWork.CommitAsync();
@@ -52,9 +46,14 @@ namespace FM.Application.Command
                     throw new Exception("User not found");
                 }
 
-                user.DisplayName = command.DisplayName;
-                user.SpotifyUserId = command.SpotifyUserId;
-                user.UserRoleId = command.UserRoleId;
+                if (!user.RowVersion.SequenceEqual(command.RowVersion))
+                {
+                    throw new Exception("The user has been modified by someone else. Please refresh and try again.");
+                }
+
+                user.UpdateDisplayName(command.DisplayName);
+                user.UpdateSpotifyUserId(command.SpotifyUserId);
+                user.UpdateUserRoleId(command.UserRoleId);
 
                 await _userRepository.UpdateUserAsync(user);
                 await _unitOfWork.CommitAsync();
@@ -75,6 +74,11 @@ namespace FM.Application.Command
                 if (user == null)
                 {
                     throw new Exception("User not found");
+                }
+
+                if (!user.RowVersion.SequenceEqual(command.RowVersion))
+                {
+                    throw new Exception("The user has been modified by someone else. Please refresh and try again.");
                 }
 
                 await _userRepository.DeleteUserAsync(command.Id, user.RowVersion);
